refactor: use devsources manifest rather than explicit test file list
