+++ docs/BUILD.bazel
@@ -21,7 +21,7 @@ skylark_doc(
         "//internal/history-server:history_server.bzl",
         "//internal/http-server:http_server.bzl",
         "//internal/jasmine_node_test:jasmine_node_test.bzl",
-        "//internal/mocha_node_test:mocha_node_test.bzl",
+        "//internal/mocha:mocha_node_test.bzl",
         "//internal/node:node.bzl",
         "//internal/node:node_repositories.bzl",
         "//internal/npm_install:npm_install.bzl",
