+++ internal/jasmine_node_test/jasmine_runner.js
@@ -1,5 +1,4 @@
 const fs = require('fs');
-const path = require('path');
 const JasmineRunner = require('jasmine/lib/jasmine');
 
 const UTF8 = {
